Epic AI to Master
Page is locked after being localized #8165

ui-tests minor change

Refactor AI related code #8149

Add WebSocket service URL for Content Operator #8121

Show translation status for component config inputs #8092

Text component translation: show status #8142

- Showing loading spinner while a Text Component is being translated

Show Juke icons for parent items of an active input #8106

AI: validation is not triggered after insert into HtmlArea #8130

Pass target language to the dialog before opening it #8124

Made Project context as the main source for language tag.

Fix highlight and scroll to the input from AI Dialog #8101

Juke icon for inputs should be togglable #8089

Change how we update data received from the AI apps #8066

Show error notification when translation failed #8098

Add content path to AI config data #8093

Save content after successful translation of all fields #8086

Better spacing in the collaboration section

Don't refresh Content Editor panel after localisation #8047

Keep failed translation indicator visible #8069

Renamed property in AIConfig

Don't hide the Juke icon when focus is inside the input #8048

updates for ui-tests

Update AI helper config #8058

Updated AiHelper to use new format.
Added null check for after expression match in DisplayName, as it may return null in some cases, which resulted in console errors on tabbing through page.

Fields update by Juke doesn't enable Save button #8045

Include text-based inputs inside X-data when translating #8009

ui-tests issue #8037

Translate content's Display Name #8035

Pass locales to translator script #8030

Update AI Translator configuration to match the new format #8023 (#8024)

* Update AI Translator configuration to match the new format #8023

 Separate container in Content Operator plugin #8014

ui-tests issue #7999

rebase 5.3

Refactor AI events #7977

Fix instructions not applied #7967

Rename isEditableByAi method

Update custom instructions handling #7884

Renamed customPrompt into instructions.
Moved instructions to Config from Data event data.
Added separate Config events for each plugin.
Made instructions update on site config update.

Update AI part to use new apps instead of Saga #7857

Update AI part to use new apps instead of Saga #7857

Renamed Saga to AI.
Added support of multiple AI plugins.

Rename translator service #7743

Assistant dialog localization #7739

- Passing available locales list to the assistant app
- Passing asset url to assistant app

AI translation: locked state look #7718

HtmlArea in an item-set is not being updated #7716

Add new Saga REST endpoints #7713

Animate text input that is being translated #7670

Request translations from AI Assistant #7671

ui-tests issue #7709

Suggest content localisation #7669

Sync CS and Saga apps #7706

Suggest content localisation #7669

Moved AIAissistant wrapper to separate instance.
Updated AiAssistant events to make it always active, even before opened.
Added checks to show confirmation dialog.

Show AI icon only for inputs in content form #7702

Fix code style.

Let form decide if it's inputs to be AI interactable #3623

Show AI icon only for inputs in content form #7702

Move Saga icon to collaboration block #7684

Add Saga icons to text-based form items #7686

Making HtmlArea discoverable for updating data-path attr #7686

Applying single field may lead to erasing data in other fields #7644

Make assistant status polling limit configurable #7640

Remove tooltip if target element was replaced #7638

Prevent AI code from being injected, if the application is stopped #7625

Make CS use Saga's REST service directly #7623

AI Assistant: pass updated Display Name to the assistant #7619

Add support for display name update from AI Assistant #7594

Support custom promp #7596

Fix access to undefined property

Enhance HtmlArea toolbar AI button #7453

HtmlArea: focusing editor should insert mention into open Assistant dialog #7455

Remove previous version of assistant from code #7452

Required for https://github.com/enonic/app-saga/issues/131

Split AiAssistant config #7398

AI Assistant: send config to the app #7395

AI Assistant: handle event with resulting data #7369

AI Assistant: send content data and schema to the assistant #7364

Add Assistant icon to the Content Wizard toolbar #7334

-fixing build

Add Assistant icon to the Content Wizard toolbar #7334

-renaming 'Saga' to 'AIAssistant

Add Assistant icon to the Content Wizard toolbar #7334

Make content form inputs (textarea and htmlarea for now) and text component fire context events to Assistant #7342

AI Assistant: define event structure to communicate with Saga #7336

fix 2 failed ui-tests

Create an Assistant MVP #7121

-More UI adjustments

Create an Assistant MVP #7121

-Invoking app-saga service to fetch AI generated response

Create an Assistant MVP #7121

-Removing redundant service
-Invoking app-saga service to fetch AI generated response

Create an Assistant MVP #7121

+++ modules/lib/src/main/resources/assets/js/app/inputtype/text/HtmlArea.ts
@@ -6,6 +6,7 @@ import {Value} from '@enonic/lib-admin-ui/data/Value';
 import {ValueType} from '@enonic/lib-admin-ui/data/ValueType';
 import {ValueTypeConverter} from '@enonic/lib-admin-ui/data/ValueTypeConverter';
 import {ValueTypes} from '@enonic/lib-admin-ui/data/ValueTypes';
+import {DefaultErrorHandler} from '@enonic/lib-admin-ui/DefaultErrorHandler';
 import {DivEl} from '@enonic/lib-admin-ui/dom/DivEl';
 import {Element, LangDirection} from '@enonic/lib-admin-ui/dom/Element';
 import {FormEl} from '@enonic/lib-admin-ui/dom/FormEl';
@@ -27,14 +28,15 @@ import 'jquery-simulate/jquery.simulate.js';
 import * as Q from 'q';
 import {ContentSummary} from '../../content/ContentSummary';
 import {ContentRequiresSaveEvent} from '../../event/ContentRequiresSaveEvent';
+import {ProjectContext} from '../../project/ProjectContext';
 import {ContentInputTypeViewContext} from '../ContentInputTypeViewContext';
-import {HTMLAreaDialogHandler} from '../ui/text/dialog/HTMLAreaDialogHandler';
+import {HTMLAreaProxy} from '../ui/text/dialog/HTMLAreaProxy';
 import {HTMLAreaHelper} from '../ui/text/HTMLAreaHelper';
 import {HtmlEditor} from '../ui/text/HtmlEditor';
 import {HtmlEditorParams} from '../ui/text/HtmlEditorParams';
 import {StylesRequest} from '../ui/text/styles/StylesRequest';
 import {HtmlAreaResizeEvent} from './HtmlAreaResizeEvent';
-import {ProjectContext} from '../../project/ProjectContext';
+import {AiConfig} from '@enonic/lib-admin-ui/form/inputtype/InputAiConfig';
 
 export class HtmlArea
     extends BaseInputTypeNotManagingAdd {
@@ -116,6 +118,14 @@ export class HtmlArea
 
             this.onOccurrenceRemoved(() => this.resetInputHeight());
         });
+
+        this.onOccurrenceValueChanged((occurrence: TextAreaWrapper, value: Value) => {
+            const editor = this.editors.find((e: HtmlAreaOccurrenceInfo) => e.textAreaWrapper === occurrence);
+
+            if (editor) {
+                editor.savedValue = value.getString();
+            }
+        });
     }
 
     getValueType(): ValueType {
@@ -139,12 +149,18 @@ export class HtmlArea
         textAreaEl.addClass(clazz);
 
         const textAreaWrapper = new TextAreaWrapper('text-area-wrapper');
+        const editor = {
+            id: editorId,
+            textAreaWrapper,
+            textAreaEl,
+            savedValue: property.hasNonNullValue() ? property.getString() : '',
+            hasStickyToolbar: false
+        };
+        this.editors.push(editor);
 
         textAreaEl.onRendered(() => {
             this.authRequest.then(() => {
-                this.initEditor(editorId, property, textAreaWrapper).then(() => {
-                    this.editors.push({id: editorId, textAreaWrapper, textAreaEl, property, hasStickyToolbar: false});
-                });
+                this.initEditor(editorId, editor.savedValue, textAreaWrapper).catch(DefaultErrorHandler.handle);
             });
         });
 
@@ -166,11 +182,13 @@ export class HtmlArea
 
     protected updateFormInputElValue(occurrence: FormInputEl, property: Property) {
         const editor: HtmlAreaOccurrenceInfo = this.getEditorInfo(occurrence.getId());
+        const newValue = property.hasNonNullValue() ? property.getString() : '';
+
         if (editor) {
-            editor.property = property;
+            editor.savedValue = newValue;
         }
 
-        this.setEditorContent(occurrence as TextArea, property);
+        this.setEditorContent(occurrence as TextArea, newValue);
     }
 
     resetInputOccurrenceElement(occurrence: Element) {
@@ -211,12 +229,13 @@ export class HtmlArea
         iframe?.setAttribute('tabindex', enable? '0' : '-1');
     }
 
-    private initEditor(id: string, property: Property, textAreaWrapper: Element): Q.Promise<HtmlEditor> {
+    private initEditor(id: string, value: string, textAreaWrapper: Element): Q.Promise<HtmlEditor> {
         const focusHandler = (e) => {
             this.resetInputHeight();
             this.notifyFocused(e);
             this.scrollToSelected(textAreaWrapper, e);
 
+            textAreaWrapper.getHTMLElement().dispatchEvent(new CustomEvent('focusin')); // for AI Assistant
             AppHelper.dispatchCustomEvent('focusin', this);
         };
 
@@ -232,6 +251,7 @@ export class HtmlArea
         const blurHandler = (e) => {
             //checking if remove occurence button clicked or not
             AppHelper.dispatchCustomEvent('focusout', this);
+            textAreaWrapper.getHTMLElement().dispatchEvent(new CustomEvent('focusout')); // for AI Assistant
 
             this.notifyBlurred(e);
         };
@@ -279,9 +299,6 @@ export class HtmlArea
             }
         };
 
-        const createDialogHandler = event => {
-            HTMLAreaDialogHandler.createAndOpenDialog(event);
-        };
 
         const editorLoadedHandler = () => {
             if (this.notInLiveEdit()) {
@@ -293,8 +310,8 @@ export class HtmlArea
             this.moveButtonToBottomBar(textAreaWrapper, '.cke_button__sourcedialog');
         };
 
-        const editorReadyHandler = () => {
-            this.setEditorContent(textAreaWrapper.findChildById(id) as TextArea, property);
+        const editorReadyHandler = (eventInfo: CKEDITOR.eventInfo) => {
+            this.setEditorContent(textAreaWrapper.findChildById(id) as TextArea, value, true);
             const editor = this.editors.find((editor: HtmlAreaOccurrenceInfo) => editor.id === id);
 
             if (editor && !this.enabled) {
@@ -312,7 +329,7 @@ export class HtmlArea
             .setEditorContainerId(id)
             .setAssetsUri(CONFIG.getString('assetsUri'))
             .setInline(false)
-            .setCreateDialogHandler(createDialogHandler)
+            .setCreateDialogHandler(HTMLAreaProxy.createAndOpenDialog)
             .setFocusHandler(focusHandler)
             .setBlurHandler(blurHandler)
             .setKeydownHandler(keydownHandler)
@@ -329,6 +346,7 @@ export class HtmlArea
             .setCustomStylesToBeUsed(true)
             .setLangDirection(this.getLangDirection())
             .setProject(this.context.project)
+            .setLabel(this.getInput().getLabel())
             .build();
 
         return HtmlEditor.create(htmlEditorParams);
@@ -468,18 +486,16 @@ export class HtmlArea
         }
     }
 
-    private setEditorContent(textArea: TextArea, property: Property): void {
+    private setEditorContent(textArea: TextArea, value: string, silent?: boolean): void {
         const editorId: string = textArea.getId();
-        const content: string = property.hasNonNullValue() ?
-                                HTMLAreaHelper.convertRenderSrcToPreviewSrc(property.getString(), this.content?.getId(),
-                                    this.context.project) : '';
 
         if (HtmlEditor.exists(editorId)) {
+            const content: string = HTMLAreaHelper.convertRenderSrcToPreviewSrc(value, this.content?.getId(), this.context.project);
             const currentData: string = HtmlEditor.getData(editorId);
             // invoke setData only if data changed
             if (content !== currentData) {
                 const afterDataSetCallback = () => {
-                    textArea.setValue(HtmlEditor.getData(editorId), true, false);
+                    textArea.setValue(HtmlEditor.getData(editorId), silent, false);
                 };
 
                 HtmlEditor.setData(editorId, content, afterDataSetCallback);
@@ -495,6 +511,7 @@ export class HtmlArea
 
     private handleEditorValueChanged(id: string, occurrence: Element) {
         const value: string = HtmlEditor.getData(id);
+
         const textAreaEl = occurrence.findChildById(id) as TextArea;
         if (textAreaEl && value !== textAreaEl.getValue()) {
             textAreaEl.setValue(value, false, true);
@@ -526,8 +543,10 @@ export class HtmlArea
         this.editors.forEach((editor) => {
             const editorId = editor.id;
 
-            this.destroyEditor(editorId);
-            this.reInitEditor(editorId);
+            if (HtmlEditor.exists(editorId)) {
+                this.destroyEditor(editorId);
+                this.reInitEditor(editorId);
+            }
         });
     }
 
@@ -582,7 +601,7 @@ export class HtmlArea
     private reInitEditor(id: string): Q.Promise<HtmlEditor> {
         const savedEditor: HtmlAreaOccurrenceInfo = this.getEditorInfo(id);
 
-        return this.initEditor(id, savedEditor.property, savedEditor.textAreaWrapper);
+        return this.initEditor(id, savedEditor.savedValue, savedEditor.textAreaWrapper);
     }
 
     private getEditorInfo(id: string): HtmlAreaOccurrenceInfo {
@@ -604,13 +623,21 @@ export class HtmlArea
         super.updateInputOccurrenceElement(textAreaEl, property, unchangedOnly);
     }
 
+    getAiConfig(): AiConfig {
+        const formContext = this.getContext().formContext;
+        return formContext ? {
+            group: formContext.getName(),
+            aiTools: formContext.getAiTools(),
+        } : super.getAiConfig();
+    }
+
 }
 
 export interface HtmlAreaOccurrenceInfo {
     id: string;
     textAreaWrapper: Element;
     textAreaEl: TextArea;
-    property: Property;
+    savedValue: string;
     hasStickyToolbar: boolean;
 }
